refactor: Move saveAddressMeta and loadAll function into Keyring class file
* rename saveAddress to saveAddressMeta

* move addPairs (from loadAll file) and rename to addAccountPairs

* stall moving `initOptions` into index.ts, since do not want to move everything at once!
