+++ packages/app-addresses/src/Editor.tsx
@@ -191,7 +191,7 @@ class Editor extends React.PureComponent<Props, State> {
       return;
     }
 
-    keyring.saveAddress(current.address(), {
+    keyring.saveAddressMeta(current.address(), {
       name: editedName,
       whenEdited: Date.now()
     });
