+++ packages/ui-keyring/src/index.ts
@@ -4,10 +4,11 @@
 
 import { KeyringPair, KeyringPair$Meta, KeyringPair$Json } from '@polkadot/util-keyring/types';
 import { SingleAddress } from './observable/types';
-import { KeyringAddress, KeyringInstance, KeyringJson$Meta, State } from './types';
+import { KeyringAddress, KeyringInstance, KeyringJson, KeyringJson$Meta, State } from './types';
 
 import store from 'store';
 import hexToU8a from '@polkadot/util/hex/toU8a';
+import isHex from '@polkadot/util/is/hex';
 import isString from '@polkadot/util/is/string';
 import createPair from '@polkadot/util-keyring/pair';
 import addressDecode from '@polkadot/util-keyring/address/decode';
@@ -17,12 +18,11 @@ import testKeyring from '@polkadot/util-keyring/testing';
 import accounts from './observable/accounts';
 import addresses from './observable/addresses';
 import development from './observable/development';
-import loadAll from './loadAll';
+import saveRecent from './address/metaRecent';
 import isAvailable from './isAvailable';
 import isPassValid from './isPassValid';
-import saveAddress from './address/meta';
-import saveRecent from './address/metaRecent';
-import { accountKey } from './defaults';
+import { accountKey, accountRegex, addressRegex } from './defaults';
+import initOptions from './options';
 
 class Keyring implements KeyringInstance {
   private state: State;
@@ -44,6 +44,19 @@ class Keyring implements KeyringInstance {
 
     this.state.keyring.addFromJson(json);
     this.state.accounts.add(json.address, json);
+  }
+
+  addAccountPairs (): void {
+    this.state.keyring
+      .getPairs()
+      .forEach((pair) => {
+        const address = pair.address();
+
+        this.state.accounts.add(address, {
+          address,
+          meta: pair.getMeta()
+        });
+      });
   }
 
   backupAccount (pair: KeyringPair, password: string): KeyringPair$Json {
@@ -136,7 +149,36 @@ class Keyring implements KeyringInstance {
   }
 
   loadAll (): void {
-    return loadAll(this.state);
+    const { accounts, addresses, keyring } = this.state;
+
+    this.addAccountPairs();
+
+    store.each((json: KeyringJson, key: string) => {
+      if (accountRegex.test(key)) {
+        if (!json.meta || !json.meta.isTesting) {
+          keyring.addFromJson(json as KeyringPair$Json);
+          accounts.add(json.address, json);
+        }
+      } else if (addressRegex.test(key)) {
+        const address = isHex(json.address)
+          ? addressEncode(hexToU8a(json.address))
+          : json.address;
+
+        // NOTE This is a fix for an older version where publicKeys instead of addresses
+        // were saved. Here we clean the old and replace with a new address-specific key
+        if (address !== json.address) {
+          json.address = address;
+
+          store.remove(key);
+          this.saveAddressMeta(address, json.meta);
+        }
+
+        addresses.add(json.address, json);
+      }
+    });
+
+    // TODO - refactor initOptions into this Keyring class file?
+    initOptions(this.state);
   }
 
   restoreAccount (json: KeyringPair$Json, password: string): KeyringPair {
@@ -178,8 +220,24 @@ class Keyring implements KeyringInstance {
     this.state.accounts.add(json.address, json);
   }
 
-  saveAddress (address: string, meta: KeyringPair$Meta): void {
-    return saveAddress(this.state, address, meta);
+  saveAddressMeta (address: string, meta: KeyringPair$Meta): void {
+    const available = this.state.addresses.subject.getValue();
+
+    const json = (available[address] && available[address].json) || {
+      address,
+      meta: {
+        isRecent: void 0,
+        whenCreated: Date.now()
+      }
+    };
+
+    Object.keys(meta).forEach((key) => {
+      json.meta[key] = meta[key];
+    });
+
+    delete json.meta.isRecent;
+
+    this.state.addresses.add(address, json);
   }
 
   saveRecent (address: string): SingleAddress {
